fix: fix simple comments that def should be fixed
