+++ routes.ini
@@ -99,19 +99,19 @@ scripts=
 /administration/new-npc=Website\Administration\Administration?new-npc
 /administration/upload=Website\Administration\Administration?mass-upload
 /administration/npcs=Website\Administration\Administration?npcs
-/administration/npcs/swap/<0>/<1>/<2>=Website\Administration\Administration?npcs?swap?<0>?<1>?<2>
-/administration/npcs/add-to-quest/<0>/<1>=Website\Administration\Administration?npcs?add-to-quest?<0>?<1>
-/administration/npcs/remove-from-quest/<0>/<1>=Website\Administration\Administration?npcs?remove-from-quest?<0>?<1>
-/administration/npcs/search=Website\Administration\Administration?npcs?search
-/administration/npcs/autocomplete=Website\Administration\Administration?npcs?autocomplete
+/administration/npcs/swap/<0>/<1>/<2>=Api\Administration\Npcs?swap?<0>?<1>?<2>
+/administration/npcs/add-to-quest/<0>/<1>=Api\Administration\Npcs?add-to-quest?<0>?<1>
+/administration/npcs/remove-from-quest/<0>/<1>=Api\Administration\Npcs?remove-from-quest?<0>?<1>
+/administration/npcs/search=Api\Administration\Npcs?search
+/administration/npcs/autocomplete=Api\Administration\Npcs?autocomplete
 /administration/npcs/manage/<0>=Website\Administration\Administration?npc?<0>?false
 /administration/npcs/manage/<0>/delete/<1>=Website\Administration\Administration?npc?<0>?false?<1>
 /administration/npcs/manage/<0>/recast/<1>=Website\Administration\Administration?npc?<0>?false?recast?<1>
 /administration/npcs/manage/<0>/archive=Website\Administration\Administration?npc?<0>?false?archive
 /administration/npcs/manage/<0>/archive-quest-recordings/<1>=Website\Administration\Administration?npc?<0>?false?archive-quest-recordings?<1>
 /administration/npcs/manage/<0>/rename=Website\Administration\Administration?npc?<0>?false?rename
-/administration/npcs/rename-quest/<0>=Website\Administration\Administration?npcs?rename-quest?<0>
-/administration/npcs/delete-quest/<0>=Website\Administration\Administration?npcs?delete-quest?<0>
+/administration/npcs/rename-quest/<0>=Api\Administration\Npcs?rename-quest?<0>
+/administration/npcs/delete-quest/<0>=Api\Administration\Npcs?delete-quest?<0>
 /administration/scripts=Website\Administration\Administration?scripts
 /logout=Website\Account\Logout
